+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.8]
+
+### Changed
+
+- Removed incrementing transaction count on compensation transactions
+- Added the ability to execute multiple use cases at once
+
 ## [2.1.7]
 
 ### Changed